Step 4. Added error handling logic.
The error field has been added to the state, which will be responsible for the caught error. We also created the onError() function, which changes the state of the error field. The updatePlanet() function has added a catch block for errors. We also changed the approach to rendering, now instead of loading, we use hasData, which, in addition to checking the loading of data, checks whether there are null objects.