Step 5. Refactor components to use class syntax and implement state management; enhance data fetching and rendering logic.
Added a state that stores an array of people. The componentDidMount() function has also been added. When the component is loaded (componentDidMount), getAllPeople() is called from the Sw api Service. Added the renderItems function for dynamic list rendering. Uses onClick to pass the id of the selected element to the parent via props.onItemSelected(id). Added conditional rendering and spinner.